colocando as view de games na tela de games

+++ app/src/main/java/com/example/testefragment/model/Game.kt
@@ -7,6 +7,6 @@ data class Game(
     var gameName: String,
     var gameImage: Drawable,
     var gameDescriptor: String,
-    var gameRating: Double,
+    var gameRating: Float,
     var consoles: List<Console>
 )
